bp_sync/pyproject.toml
[tool.poetry]
name = "bp_sync"
version = "0.1.0"
description = "bitrix production sync"
authors = ["VlNagibin"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
orjson = "^3.10.15"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
uvicorn = "^0.34.0"
SQLAlchemy = {version = "^2.0.37", extras = ["asyncio"]}
Alembic = "^1.14.0"
sqladmin = "^0.20.1"
pandas = "^2.2.3"
redis = "^6.2.0"
cryptography = "^45.0.4"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
itsdangerous = "^2.2.0"
aio-pika = "^9.5.7"
openpyxl = "^3.1.5"
types-pytz = "^2025.2.0.20250809"
requests = "^2.32.5"
types-requests = "^2.32.4.20250913"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.2"
mypy = "^1.15.0"
isort = "^6.0.1"
pre-commit = "^4.1.0"

[tool.poetry.group.tests.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-subtests = "^0.11.0"
requests = "^2.32.3"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
force_grid_wrap = 0
use_parentheses = true
force_sort_within_sections = false
multi_line_output = 3
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = 3.12
follow_imports = "skip"

[tool.pytest.ini_options]
# asyncio_mode=auto
norecursedirs = "env/* venv/*"
addopts = [
    "-vv",
    "-p",
    "no:cacheprovider",
    "--strict-markers",
]
testpaths = "tests/"
python_files = "test_*.py"
markers = [
    "test",
    "core"
]
